FRONT-9 > доделал карточку редактирования лицензии + начал карточку редактирования бита

+++ server/models/models.js
@@ -53,10 +53,10 @@ const License = sequelize.define('license', {
   type: { type: DataTypes.STRING, allowNull: false },
   name: { type: DataTypes.STRING, allowNull: false },
   price: { type: DataTypes.INTEGER, defaultValue: 1000, allowNull: false },
-  count_streams: { type: DataTypes.INTEGER, defaultValue: 50000, allowNull: false },
-  count_copies: { type: DataTypes.INTEGER, defaultValue: 5000, allowNull: false },
-  count_video_streams: { type: DataTypes.INTEGER, defaultValue: 10000, allowNull: false },
-  count_performances: { type: DataTypes.INTEGER, defaultValue: 5, allowNull: false },
+  count_streams: { type: DataTypes.BIGINT, defaultValue: 50000, allowNull: false },
+  count_copies: { type: DataTypes.BIGINT, defaultValue: 5000, allowNull: false },
+  count_video_streams: { type: DataTypes.BIGINT, defaultValue: 10000, allowNull: false },
+  count_performances: { type: DataTypes.BIGINT, defaultValue: 5, allowNull: false },
   is_visible: { type: DataTypes.BOOLEAN, defaultValue: true, allowNull: false }
 });
